requirements.txt
boto3
django
django-environ
djangorestframework
django-cors-headers
django-storages
drf-yasg
pillow